+++ habapp_rules/system/summer_winter.py
@@ -1,5 +1,4 @@
 """Rule to detect whether it is summer or winter."""
-import collections
 import datetime
 import logging
 import statistics
@@ -29,7 +28,7 @@ class SummerWinter(HABApp.Rule):
 		:param temperature_threshold: threshold weighted temperature for summer
 		:param last_check_name: Name of last check item. OpenHAB-Type must be DateTime item
 		"""
-		super().__init__()
+		HABApp.Rule.__init__(self)
 		self._instance_logger = habapp_rules.core.logger.InstanceLogger(LOGGER, summer_name)
 
 		# set class variables
@@ -46,6 +45,8 @@ class SummerWinter(HABApp.Rule):
 		# run at init and every day at 23:00
 		self.run.soon(self._cb_update_summer)
 		self.run.on_every_day(datetime.time(23), self._cb_update_summer)
+
+		LOGGER.debug("Init of Summer / Winter successful")
 
 	def __get_weighted_mean(self, days_in_past: int) -> float:
 		"""Get weighted mean temperature.
@@ -64,11 +65,10 @@ class SummerWinter(HABApp.Rule):
 		for hour in [7, 14, 22]:
 			start_time = datetime.datetime(self.__now.year, self.__now.month, self.__now.day, hour, 0) - datetime.timedelta(days=day_offset + days_in_past)
 			end_time = start_time + datetime.timedelta(hours=1)
-			values = self._outside_temp_item.get_persistence_data(persistence=self._persistence_service, start_time=start_time, end_time=end_time)
-			values_sorted = collections.OrderedDict(sorted(values.get_data().items()))
-			if not values_sorted:
-				raise SummerWinterException(f"No data for days_in_past = {days_in_past} and hour = {hour}")
-			temperature_values.append(list(values_sorted.values())[0])
+			persistence_data = self._outside_temp_item.get_persistence_data(persistence=self._persistence_service, start_time=start_time, end_time=end_time)
+			if not persistence_data.data:
+				raise SummerWinterException(f"No data for {start_time}")
+			temperature_values.append(next(iter(persistence_data.data.values())))
 		return (sum(temperature_values) + temperature_values[2]) / 4
 
 	def __is_summer(self) -> bool:
@@ -85,8 +85,8 @@ class SummerWinter(HABApp.Rule):
 			except SummerWinterException:
 				self._instance_logger.warning(f"Could not get mean value of day -{day}")
 
-		if len(values) <= self._days * 0.5:
-			raise SummerWinterException(f"Not enough values to detect summer/winter. Expected: {self._days} | actual: {len(values)}")
+		if not values:
+			raise SummerWinterException("Not enough data to detect summer/winter")
 
 		is_summer = self._hysteresis_switch.get_output(mean_value := statistics.mean(values))
 		self._instance_logger.debug(f"Check Summer/Winter. values = {values} | mean = {mean_value} | summer = {is_summer}")
@@ -97,7 +97,7 @@ class SummerWinter(HABApp.Rule):
 		try:
 			is_summer = self.__is_summer()
 		except SummerWinterException:
-			self._instance_logger.exception("Could not get summer / winter")
+			self._instance_logger.error("Could not get summer / winter")
 			return
 
 		# get target state of summer